improve crud

+++ app/Http/Controllers/PostController.php
@@ -7,32 +7,48 @@ use Illuminate\Http\Request;
 
 class PostController extends Controller
 {
-    
+    public function __construct(Post $post){
+        $this->post = $post;
+    }
     public function index()
     {
-        $posts = Post::all();
+        // $posts = Post::all(); Método estático
+        $posts = $this->post->all();
         return $posts;
     }
     
     public function store(Request $request)
     {
-        $new_post = Post::create($request->all());
+        $new_post = $this->post->create($request->all());
         return $new_post;
     }
 
-    public function show(Post $post)
+    // Recebe um inteiro id e pesquiso ele entre meus itens armazenados
+    public function show($id)
     {
+        $post = $this->post->find($id);
+        if ($post === null){
+            return ['error:'=>'Route not found'];
+        }
         return $post;
     }
    
-    public function update(Request $request, Post $post)
+    public function update(Request $request, $id)
     {
+        $post = $this->post->find($id);
+        if ($post === null){
+            return ['error:'=>'Route not found. update failed'];
+        }
         $post->update($request->all());
         return $post;
     }
 
-    public function destroy(Post $post)
+    public function destroy($id)
     {
+        $post = $this->post->find($id);
+        if ($post === null){
+            return ['error:'=>'Route not found. delete failed'];
+        }
         $post->delete();
         return ['msg:' => 'The post has been deleted succefully'];
     }
